Fix symbol name in ticker reports
When writing Python code, I had assumed a for-loop created a new scoped
block, and that variables defined inside of it where only inside of it.
It turns out that is not true, so in an async world where you are using
a variable defined as part of the for-loop, the value is liable to
change out from under you.

A good hack is to define a default argument for the function you are
defining inside the loop with the same variable, as the function you
define has its own scope block.

Also fixed a couple instances of this bug in the core.