Fix failure when non-UTF-8 characters received
Closes #176.

The bug here is two part:

1. Cardinal assumes incoming messages will be UTF-8 encoded, and raises
   an exception if a non-UTF-8 characater appears in the message.
2. `twisted.words.protocol.irc.IRCClient` has the same behavior.

The IRC RFC does not specify a message encoding:

> 2.2 Character codes
>
>    No specific character set is specified. The protocol is based on a
>    set of codes which are composed of eight (8) bits, making up an
>    octet.  Each message may be composed of any number of these octets;
>    however, some octet values are used for control codes, which act as
>    message delimiters.
>
>    Regardless of being an 8-bit protocol, the delimiters and keywords
>    are such that protocol is mostly usable from US-ASCII terminal and a
>    telnet connection.

Still, for my use case, and probably the use case of most who would use
Cardinal, UTF-8 is as close of a character encoding as we will get. The
message that caused the issue creating this bug was mostly UTF-8, with
some control characters included.

The solution implemented in this diff is to first decode the message
into a UTF-8 string, replacing any non-UTF-8 characters with the Unicode
replacement character. Then, we re-encode the message as bytes and pass
it back to Twisted with the non-UTF-8 characters removed.