Remove mIRC formatting when responding to .seen
Fixes #182. Channels with `+c` set on them would reject messages from
the `.seen` plugin as many of them included the `0x03` character, to
reset any remaining colors on the message. There were a couple problems
with this approach:

1. It wasn't very intelligent - this only worked to fix colors which
   hadn't been terminated, but did not help in the event of bold,
   underlined, italic, strikethrough, or monospace text.
2. This control code was being sent regardless of whether it was
   necessary. More ideally, we'd check if the message actually needed to
   be reset.
3. As noted in the original ticket, this broke the plugin for channels
   with `+c` (block messages with formatting) set.

If we intend to bring back formatting in the future, we should aim to
satisfy the following conditions:

1. Continue to strip formatting for `+c` channels.
2. Only terminate unterminated formatting codes.
3. Look for more than just colors.