Fix a bug with plugin reloading after Exception in constructor
There are many scenarios where a plugin can be reloaded:

1. When a plugin is currently loaded.
2. After a plugin failed to load due to an ImportError.
3. After a plugin failed to load due to an exception in its constructor.

In scenario #1, we have a module object in scope we can call `reload()`
on. In scenario #2, we don't need said module object as it was never
loaded. In scenario #3, we loaded the module previously but don't have
it in scope anymore to call `reload()` on.

This commit fixes the problem by creating a `_module_cache` which lives
separately from `self.plugins`. It contains all successfully imported
plugin modules, keyed by their plugin name.