Add first pass at cryptocurrency plugin
Borrows relaying code from `ticker` but I am trying out a new method of
attaching the regex secondarily to a command. This has nice properties
in that it behaves similarly to normal (e.g. with help, correct prefix,
etc.) but also supports relayed bots.

Eventually, it'd be nice to move some of this into the core - we'll
probably need to specify another param to `@command()` like
`supports_relayed=True` to avoid accidentally giving incorrect
privileges to a user on the other side of the relay.