Fix urls plugin reloading
I accidentally removed the close() method that de-registers the
urls.detection event in the plugin, when URL shortening was added. This
puts it back, and also updates the core to avoid de-registering
callbacks. All the way back when I first created the event system, I
made sure that you could register callbacks before the event, because
plugins can load in any order. I didn't consider that I was creating a
similar problem by removing the callbacks when the event was removed.